refactor: 도메인 검증 로직을 도메인 서비스로 이동
