+++ src/main/java/com/gdschongik/gdsc/domain/recruitment/application/AdminRecruitmentService.java
@@ -9,6 +9,7 @@ import com.gdschongik.gdsc.domain.recruitment.dao.RecruitmentRepository;
 import com.gdschongik.gdsc.domain.recruitment.dao.RecruitmentRoundRepository;
 import com.gdschongik.gdsc.domain.recruitment.domain.Recruitment;
 import com.gdschongik.gdsc.domain.recruitment.domain.RecruitmentRound;
+import com.gdschongik.gdsc.domain.recruitment.domain.RecruitmentRoundValidator;
 import com.gdschongik.gdsc.domain.recruitment.domain.RoundType;
 import com.gdschongik.gdsc.domain.recruitment.dto.request.RecruitmentCreateRequest;
 import com.gdschongik.gdsc.domain.recruitment.dto.request.RecruitmentRoundCreateRequest;
@@ -16,7 +17,6 @@ import com.gdschongik.gdsc.domain.recruitment.dto.request.RecruitmentRoundUpdate
 import com.gdschongik.gdsc.domain.recruitment.dto.response.AdminRecruitmentResponse;
 import com.gdschongik.gdsc.global.exception.CustomException;
 import java.time.LocalDateTime;
-import java.time.Month;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -31,6 +31,7 @@ public class AdminRecruitmentService {
 
     private final RecruitmentRepository recruitmentRepository;
     private final RecruitmentRoundRepository recruitmentRoundRepository;
+    private final RecruitmentRoundValidator recruitmentRoundValidator;
 
     @Transactional
     public void createRecruitment(RecruitmentCreateRequest request) {}
@@ -42,11 +43,9 @@ public class AdminRecruitmentService {
 
     @Transactional
     public void createRecruitmentRound(RecruitmentRoundCreateRequest request) {
-        validatePeriodMatchesAcademicYear(request.startDate(), request.endDate(), request.academicYear());
-        validatePeriodMatchesSemesterType(request.startDate(), request.endDate(), request.semesterType());
-        validatePeriodOverlap(request.academicYear(), request.semesterType(), request.startDate(), request.endDate());
-        validatePeriodWithinTwoWeeks(
+        recruitmentRoundValidator.validateRecruitmentRoundCreate(
                 request.startDate(), request.endDate(), request.academicYear(), request.semesterType());
+        validatePeriodOverlap(request.academicYear(), request.semesterType(), request.startDate(), request.endDate());
         validateRound(request.academicYear(), request.semesterType(), request.roundType());
 
         Recruitment recruitment = recruitmentRepository
@@ -82,69 +81,6 @@ public class AdminRecruitmentService {
         }
 
         recruitmentRounds.forEach(RecruitmentRound::validatePeriodNotStarted);
-    }
-
-    // TODO validateRegularRequirement처럼 로직 변경
-    private void validatePeriodMatchesAcademicYear(
-            LocalDateTime startDate, LocalDateTime endDate, Integer academicYear) {
-        if (academicYear.equals(startDate.getYear()) && academicYear.equals(endDate.getYear())) {
-            return;
-        }
-
-        throw new CustomException(RECRUITMENT_PERIOD_MISMATCH_ACADEMIC_YEAR);
-    }
-
-    // TODO validateRegularRequirement처럼 로직 변경
-    private void validatePeriodMatchesSemesterType(
-            LocalDateTime startDate, LocalDateTime endDate, SemesterType semesterType) {
-        if (getSemesterTypeByStartDateOrEndDate(startDate).equals(semesterType)
-                && getSemesterTypeByStartDateOrEndDate(endDate).equals(semesterType)) {
-            return;
-        }
-
-        throw new CustomException(RECRUITMENT_PERIOD_MISMATCH_SEMESTER_TYPE);
-    }
-
-    private SemesterType getSemesterTypeByStartDateOrEndDate(LocalDateTime dateTime) {
-        int year = dateTime.getYear();
-        LocalDateTime firstSemesterStartDate = LocalDateTime.of(
-                year, FIRST.getStartDate().getMonth(), FIRST.getStartDate().getDayOfMonth(), 0, 0);
-        LocalDateTime secondSemesterStartDate = LocalDateTime.of(
-                year, SECOND.getStartDate().getMonth(), SECOND.getStartDate().getDayOfMonth(), 0, 0);
-
-        /*
-        개강일 기준으로 2주 전까지는 같은 학기로 간주한다.
-         */
-        if (dateTime.isAfter(firstSemesterStartDate.minusWeeks(PRE_SEMESTER_TERM))
-                && dateTime.getMonthValue() < Month.JULY.getValue()) {
-            return FIRST;
-        }
-
-        if (dateTime.isAfter(secondSemesterStartDate.minusWeeks(PRE_SEMESTER_TERM))) {
-            return SECOND;
-        }
-
-        throw new CustomException(RECRUITMENT_PERIOD_SEMESTER_TYPE_UNMAPPED);
-    }
-
-    private void validatePeriodWithinTwoWeeks(
-            LocalDateTime startDate, LocalDateTime endDate, Integer academicYear, SemesterType semesterType) {
-        LocalDateTime semesterStartDate = LocalDateTime.of(
-                academicYear,
-                semesterType.getStartDate().getMonth(),
-                semesterType.getStartDate().getDayOfMonth(),
-                0,
-                0);
-
-        if (semesterStartDate.minusWeeks(PRE_SEMESTER_TERM).isAfter(startDate)
-                || semesterStartDate.plusWeeks(PRE_SEMESTER_TERM).isBefore(startDate)) {
-            throw new CustomException(RECRUITMENT_PERIOD_NOT_WITHIN_TWO_WEEKS);
-        }
-
-        if (semesterStartDate.minusWeeks(PRE_SEMESTER_TERM).isAfter(endDate)
-                || semesterStartDate.plusWeeks(PRE_SEMESTER_TERM).isBefore(endDate)) {
-            throw new CustomException(RECRUITMENT_PERIOD_NOT_WITHIN_TWO_WEEKS);
-        }
     }
 
     // 새로 생성하는 경우
