+++ src/test/java/com/gdschongik/gdsc/domain/recruitment/application/AdminRecruitmentServiceTest.java
@@ -4,7 +4,6 @@ import static com.gdschongik.gdsc.global.common.constant.RecruitmentConstant.*;
 import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 import static org.assertj.core.api.Assertions.*;
 
-import com.gdschongik.gdsc.domain.common.model.SemesterType;
 import com.gdschongik.gdsc.domain.recruitment.domain.RoundType;
 import com.gdschongik.gdsc.domain.recruitment.dto.request.RecruitmentRoundCreateRequest;
 import com.gdschongik.gdsc.global.exception.CustomException;
@@ -139,30 +138,6 @@ class AdminRecruitmentServiceTest extends IntegrationTest {
     class 모집회차_생성시 {
 
         @Test
-        void 모집_시작일과_종료일의_연도가_입력된_학년도와_다르다면_실패한다() {
-            // given
-            RecruitmentRoundCreateRequest request = new RecruitmentRoundCreateRequest(
-                    2025, SEMESTER_TYPE, RECRUITMENT_NAME, START_DATE, END_DATE, ROUND_TYPE);
-
-            // when & then
-            assertThatThrownBy(() -> adminRecruitmentService.createRecruitmentRound(request))
-                    .isInstanceOf(CustomException.class)
-                    .hasMessage(RECRUITMENT_PERIOD_MISMATCH_ACADEMIC_YEAR.getMessage());
-        }
-
-        @Test
-        void 학기_시작일과_종료일의_학기가_입력된_학기와_다르다면_실패한다() {
-            // given
-            RecruitmentRoundCreateRequest request = new RecruitmentRoundCreateRequest(
-                    ACADEMIC_YEAR, SemesterType.SECOND, RECRUITMENT_NAME, START_DATE, END_DATE, ROUND_TYPE);
-
-            // when & then
-            assertThatThrownBy(() -> adminRecruitmentService.createRecruitmentRound(request))
-                    .isInstanceOf(CustomException.class)
-                    .hasMessage(RECRUITMENT_PERIOD_MISMATCH_SEMESTER_TYPE.getMessage());
-        }
-
-        @Test
         void 학년도_학기_차수가_모두_중복되면_실패한다() {
             // given
             createRecruitmentRound(
@@ -206,22 +181,5 @@ class AdminRecruitmentServiceTest extends IntegrationTest {
                     .isInstanceOf(CustomException.class)
                     .hasMessage(PERIOD_OVERLAP.getMessage());
         }
-    }
-
-    @Test
-    void 모집_시작일과_종료일이_학기_시작일로부터_2주_이내에_있지_않다면_실패한다() {
-        // given
-        RecruitmentRoundCreateRequest request = new RecruitmentRoundCreateRequest(
-                ACADEMIC_YEAR,
-                SEMESTER_TYPE,
-                RECRUITMENT_NAME,
-                START_DATE,
-                LocalDateTime.of(2024, 4, 10, 0, 0),
-                RoundType.SECOND);
-
-        // when & then
-        assertThatThrownBy(() -> adminRecruitmentService.createRecruitmentRound(request))
-                .isInstanceOf(CustomException.class)
-                .hasMessage(RECRUITMENT_PERIOD_NOT_WITHIN_TWO_WEEKS.getMessage());
     }
 }
